apps/api/requirements-sprint1.txt
# ===========================================
# C2PRO Backend - Dependencies (Sprint 1)
# ===========================================
# Versiones compatibles con Python 3.13 en Windows

# Framework (versiones más recientes con wheels para Python 3.13)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Database
sqlalchemy>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0

# Security
PyJWT[crypto]>=2.8.0
passlib[bcrypt]>=1.7.4
cryptography>=44.0.0

# Observability (básico)
structlog>=24.4.0

# Utilities
python-dotenv>=1.0.1
asgi-lifespan==2.1.0
